+++ src/main/java/br/com/academico/endereco/EnderecoService.java
@@ -12,15 +12,15 @@ public class EnderecoService implements IEnderecoService {
     
     public List<Endereco> listar(){
         List<Endereco> listaEnderecos = new ArrayList<Endereco>();
-		listaEnderecos.add(new Endereco(55555, "Rua da feira", "Centro", "Aracaju", "Sergipe"));
-		listaEnderecos.add(new Endereco(66666, "Rua da igreja", "Atalaia", "Aracaju", "Sergipe"));
+		listaEnderecos.add(new Endereco(55555L, "Rua da feira", "Centro", "Aracaju", "Sergipe"));
+		listaEnderecos.add(new Endereco(66666L, "Rua da igreja", "Atalaia", "Aracaju", "Sergipe"));
         return listaEnderecos;
     }
 
-    public Endereco recuperar(int id) {
+    public Endereco recuperar(Long id) {
         Endereco endereco;
         if (id != 999){
-            endereco = new Endereco(88888, "Rua B", "Centro", "Tobias Barreto", "Sergipe");
+            endereco = new Endereco(88888L, "Rua B", "Centro", "Tobias Barreto", "Sergipe");
             endereco.setId(id);
         }
         else {
@@ -29,9 +29,9 @@ public class EnderecoService implements IEnderecoService {
         return endereco;
     }
 
-    public int criar(Endereco endereco) {
+    public Long criar(Endereco endereco) {
         if (endereco.getCEP() != 88888) {
-            endereco.setId(200);
+            endereco.setId(200L);
         }
         else {
             throw new CEPEnderecoInvalidoException();
@@ -39,7 +39,7 @@ public class EnderecoService implements IEnderecoService {
         return endereco.getId();
     }
 
-    public Endereco atualizar(int id, Endereco endereco) {
+    public Endereco atualizar(Long id, Endereco endereco) {
         if (id != 999) {
             endereco.setId(id);
             endereco.setRua("Rua Nova");
@@ -50,11 +50,11 @@ public class EnderecoService implements IEnderecoService {
         return endereco;
     }
 
-    public int deletar(int id) {
+    public Long deletar(Long id) {
         return id;
     }
 
-    public Endereco mudarStatus(int id, StatusEndereco status) {
+    public Endereco mudarStatus(Long id, StatusEndereco status) {
         Endereco endereco = new Endereco();
         endereco.setId(id);
         endereco.setStatus(status);
